requirements.txt
python-docx>=0.8.11
sentence-transformers>=2.2.2
chromadb>=0.4.15
llama-cpp-python>=0.2.23
python-telegram-bot>=22.1
tqdm>=4.66.1